# web/api/openapi.yaml
openapi: 3.0.3
info:
  title: 打印服务器 SN/DeviceName 管理中心 API
  description: |
    与 API.md 为同一份契约(唯一通信标准)。code=0 成功;业务错误 HTTP 200 + 业务码。
  version: "1.0"
servers:
  - url: /api/v1
tags:
  - name: 账号
  - name: 商户
  - name: 商户设备
  - name: 设备
  - name: 入库
  - name: 设备端
  - name: 总览

paths:
  /login:
    post:
      tags: [账号]
      summary: 登录
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginReq" }
      responses:
        "200":
          description: 成功 / 业务错误(1003 用户名或密码错误、1004 账号已停用)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResp" }

  /me:
    get:
      tags: [账号]
      summary: 当前账号
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/UserData" }

  /me/password:
    put:
      tags: [账号]
      summary: 修改自己密码
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordReq" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }

  /accounts:
    get:
      tags: [账号]
      summary: 账号列表(admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20 } }
        - { name: keyword, in: query, schema: { type: string } }
      responses:
        "200":
          description: data={total, items:[User]}
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageUsers" }
    post:
      tags: [账号]
      summary: 新建账号(admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAccountReq" }
      responses:
        "200": { $ref: "#/components/responses/UserData" }

  /accounts/{id}:
    put:
      tags: [账号]
      summary: 编辑账号(admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAccountReq" }
      responses:
        "200": { $ref: "#/components/responses/UserData" }

  /accounts/{id}/reset-password:
    put:
      tags: [账号]
      summary: 重置密码(admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [newPassword], properties: { newPassword: { type: string, minLength: 6, maxLength: 64 } } }
      responses:
        "200": { $ref: "#/components/responses/Ok" }

  /merchants:
    get:
      tags: [商户]
      summary: 商户列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: data={items:[Merchant]}
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MerchantListData" }
    post:
      tags: [商户]
      summary: 新增商户(新增完成即入库)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMerchantReq" }
      responses:
        "200":
          description: 成功 / 2004 商户号(长/短)已存在
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MerchantData" }

  /merchants/{id}:
    get:
      tags: [商户]
      summary: 商户详情
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { $ref: "#/components/responses/MerchantData" }
    delete:
      tags: [商户]
      summary: 删除商户(admin;有已绑定设备拒绝,未绑定设备回库存)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }

  /merchants/{id}/allocate:
    post:
      tags: [商户设备]
      summary: 分配设备(从库存提取,商户独占30天)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [count], properties: { count: { type: integer, minimum: 1 } } }
      responses:
        "200":
          description: 成功 / 3006 库存不足(整批不分配)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AllocateData" }

  /merchants/{id}/devices:
    get:
      tags: [商户设备]
      summary: 名下设备列表
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: data={items:[Device]}
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceListData" }

  /merchants/{id}/devices/{name}/reclaim:
    post:
      tags: [商户设备]
      summary: 收回设备(仅已分配未绑定)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/Ok" }

  /devices:
    get:
      tags: [设备]
      summary: 设备总表(只读)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20 } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: merchantId, in: query, schema: { type: integer } }
        - { name: state, in: query, schema: { type: string, enum: [inventory, allocated, bound] } }
      responses:
        "200": { $ref: "#/components/responses/DevicePage" }

  /devices/{name}:
    get:
      tags: [设备]
      summary: 设备详情
      security: [{ bearerAuth: [] }]
      parameters: [{ name: name, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: 成功 / 3002 设备不存在
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceDetailData" }

  /devices/{name}/unbind:
    post:
      tags: [设备]
      summary: 解绑(admin,换硬件/回收场景)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: name, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }

  /devices/batch-unbind:
    post:
      tags: [设备]
      summary: 批量解绑(admin;逐台解绑,返回成功与跳过明细)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names: { type: array, minItems: 1, maxItems: 200, items: { type: string } }
      responses:
        "200":
          description: data={unbound:[...],skipped:{name:reason}} / 2001 参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      unbound: { type: array, items: { type: string } }
                      skipped: { type: object, additionalProperties: { type: string } }

  /devices/batch-reclaim:
    post:
      tags: [设备]
      summary: 批量回收(admin;已分配未激活收回库存,返回成功与跳过明细)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names: { type: array, minItems: 1, maxItems: 200, items: { type: string } }
      responses:
        "200":
          description: data={reclaimed:[...],skipped:{name:reason}} / 2001 参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      reclaimed: { type: array, items: { type: string } }
                      skipped: { type: object, additionalProperties: { type: string } }

  /inventory/import:
    post:
      tags: [入库]
      summary: 导入设备 CSV(admin;整批原子,失败零写入)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary, description: UTF-8 CSV,表头 DeviceName,DeviceSecret,ProductKey }
      responses:
        "200":
          description: 成功 / 4002 模板格式错误 / 4003 文件内DeviceName重复 / 4004 与库存DeviceName重复
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportData" }

  /inventory:
    get:
      tags: [入库]
      summary: 库存列表(admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 50 } }
      responses:
        "200": { $ref: "#/components/responses/DevicePage" }

  /device/lookup:
    post:
      tags: [设备端]
      summary: 查询可认领设备/重装恢复探测(设备端,无JWT)
      parameters: [{ name: X-Device-Name, in: header, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceLookupReq" }
      responses:
        "200":
          description: 成功 / 2003 商户号不存在 / 3004 指纹无效 / 3001 暂无可认领设备
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceLookupData" }

  /device/bind:
    post:
      tags: [设备端]
      summary: 绑定/恢复(幂等兼心跳;返回密钥供客户端写本地MQTT配置)
      parameters: [{ name: X-Device-Name, in: header, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceBindReq" }
      responses:
        "200":
          description: 成功 / 3002 不属于该商户 / 3003 已绑其他设备 / 3004 指纹无效 / 3005 抢注失败
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceBindData" }

  /dashboard:
    get:
      tags: [总览]
      summary: 仪表盘统计
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: data={merchantCount, inventoryCount, allocatedCount, boundCount, loginStats:[{date,success,failed}]}
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DashboardData" }

components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }

  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer } }

  responses:
    Ok:
      description: 成功(或业务错误码,见 API.md 错误码总表)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
    UserData:
      description: data=User
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/Envelope" }
              - type: object
                properties: { data: { $ref: "#/components/schemas/User" } }
    MerchantData:
      description: data=Merchant
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/Envelope" }
              - type: object
                properties: { data: { $ref: "#/components/schemas/Merchant" } }
    DevicePage:
      description: data={total, items:[Device]}
      content:
        application/json:
          schema: { $ref: "#/components/schemas/DeviceListData" }

  schemas:
    Envelope:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data: { nullable: true }

    LoginReq:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }

    LoginResp:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                token: { type: string }
                user: { $ref: "#/components/schemas/User" }

    ChangePasswordReq:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword: { type: string }
        newPassword: { type: string, minLength: 6, maxLength: 64 }

    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [admin, operator] }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }

    PageUsers:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                total: { type: integer }
                items: { type: array, items: { $ref: "#/components/schemas/User" } }

    CreateAccountReq:
      type: object
      required: [username, password, role]
      properties:
        username: { type: string, pattern: "^[A-Za-z0-9_]{2,32}$" }
        password: { type: string, minLength: 6, maxLength: 64 }
        displayName: { type: string }
        role: { type: string, enum: [admin, operator] }

    UpdateAccountReq:
      type: object
      properties:
        displayName: { type: string }
        role: { type: string, enum: [admin, operator] }
        enabled: { type: boolean }

    Merchant:
      type: object
      properties:
        id: { type: integer }
        merchantNoLong: { type: string }
        merchantNoShort: { type: string }
        name: { type: string }
        contactPhone: { type: string }
        address: { type: string }
        remark: { type: string }
        allocatedCount: { type: integer }
        boundCount: { type: integer }

    MerchantListData:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/Merchant" } }

    CreateMerchantReq:
      type: object
      required: [merchantNoLong, merchantNoShort, name]
      properties:
        merchantNoLong: { type: string, minLength: 1, maxLength: 32 }
        merchantNoShort: { type: string, minLength: 1, maxLength: 32 }
        name: { type: string, minLength: 1, maxLength: 64 }
        contactPhone: { type: string }
        address: { type: string }
        remark: { type: string }

    AllocateData:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                allocated: { type: array, items: { type: string }, description: 分配到的 DeviceName 列表 }

    Device:
      type: object
      properties:
        name: { type: string, description: DeviceName(设备SN) }
        deviceSecret: { type: string, description: 设备密钥 }
        productKey: { type: string, description: 通信密钥 }
        merchantId: { type: integer, nullable: true }
        merchantNoLong: { type: string, nullable: true }
        merchantNoShort: { type: string, nullable: true }
        merchantName: { type: string, nullable: true }
        state: { type: string, enum: [inventory, allocated, bound] }
        allocatedAt: { type: string, format: date-time, nullable: true }
        allocatedLeftDays: { type: integer, description: 独占期剩余天数(未激活0~30,否则0) }
        boundAt: { type: string, format: date-time, nullable: true }
        lastSeenAt: { type: string, format: date-time, nullable: true }
        online: { type: boolean, description: lastSeenAt 距今<120s }
        osType: { type: string, enum: [win, android], nullable: true, description: 客户端平台 }
        appVersion: { type: string, nullable: true, description: 客户端版本号 }

    DeviceListData:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                total: { type: integer }
                items: { type: array, items: { $ref: "#/components/schemas/Device" } }

    DeviceDetailData:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              allOf:
                - { $ref: "#/components/schemas/Device" }
                - type: object
                  properties:
                    merchant: { type: object, nullable: true }
                    fingerprint: { $ref: "#/components/schemas/Fingerprint" }

    ImportData:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                imported: { type: integer }
                total: { type: integer }

    Fingerprint:
      type: object
      description: 身份字段4个参与摘要;boardSerial/diskSerials 至少一项非空;不收网卡MAC;diskSerials仅系统盘
      properties:
        osType: { type: string, enum: [win, android] }
        boardSerial: { type: string }
        cpuId: { type: string }
        diskSerials: { type: array, items: { type: string }, description: 系统盘序列号数组 }
        osBuild: { type: string, nullable: true }
        appVersion: { type: string, nullable: true }
        deviceModel: { type: string, nullable: true }

    DeviceInfo:
      type: object
      required: [osType]
      properties:
        osType: { type: string, enum: [win, android] }
        appVersion: { type: string }
        osBuild: { type: string }
        deviceModel: { type: string }

    DeviceLookupReq:
      type: object
      required: [merchantNo, fingerprint]
      properties:
        merchantNo: { type: string, description: 长或短商户号 }
        fingerprint: { $ref: "#/components/schemas/Fingerprint" }
        deviceInfo: { $ref: "#/components/schemas/DeviceInfo" }

    DeviceLookupData:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                merchant: { type: object, nullable: true }
                boundDevice:
                  type: object
                  nullable: true
                  description: 该指纹已绑定设备(重装恢复),名称+密钥供客户端回填配置
                  properties:
                    name: { type: string }
                    productKey: { type: string }
                    deviceSecret: { type: string }
                availableDevices:
                  type: array
                  description: 该商户名下可认领设备(名称+密钥);绑定确认在客户端「保存设置」后的 bind 调用
                  items:
                    type: object
                    properties:
                      name: { type: string }
                      productKey: { type: string }
                      deviceSecret: { type: string }
                total: { type: integer }
                boundCount: { type: integer }

    DeviceBindReq:
      type: object
      required: [merchantNo, deviceName, fingerprint]
      properties:
        merchantNo: { type: string }
        deviceName: { type: string }
        fingerprint: { $ref: "#/components/schemas/Fingerprint" }
        deviceInfo: { $ref: "#/components/schemas/DeviceInfo" }

    DeviceBindData:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                deviceName: { type: string }
                deviceSecret: { type: string, description: 供客户端写入本地MQTT配置 }
                productKey: { type: string, description: 供客户端写入本地MQTT配置 }
                status: { type: string, example: bound }

    DashboardData:
      allOf:
        - { $ref: "#/components/schemas/Envelope" }
        - type: object
          properties:
            data:
              type: object
              properties:
                merchantCount: { type: integer }
                inventoryCount: { type: integer }
                allocatedCount: { type: integer }
                boundCount: { type: integer }
                loginStats:
                  type: array
                  items:
                    type: object
                    properties:
                      date: { type: string, format: date }
                      success: { type: integer }
                      failed: { type: integer }